face_identification/requirements.txt
face_recognition
flask
mysql-connector-python
numpy
# git+https://github.com/ageitgey/face_recognition_models
